Pulled some setup in the regular contributions checkout out into a helper.
